LPS-104378 Simplifying naming
